requirements.txt
numpy==1.19.5
tensorflow_gpu==2.3.0
Pillow==8.3.1
tensorflow==2.6.0
